requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
pandas==2.0.3
numpy>1.24.3
joblib==1.3.2
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
catboost==1.2.8
click==8.1.7
blinker==1.6.2
python-dotenv==1.1.1
openai==1.51.0
google-generativeai==0.8.3
Pillow==10.4.0
markdown==3.7
psycopg2-binary==2.9.7
Flask-Login==0.6.3
requests==2.31.0
psycopg2-binary
cloud-sql-python-connector[pg8000]
bcrypt